Factor out CSV re-coder
