+++ messytables/commas.py
@@ -1,76 +1,13 @@
 import csv
-import codecs
-import chardet
 
-from six import text_type, binary_type, PY2
+from six import text_type, PY2
 
-from messytables.buffered import seekable_stream
+from messytables.buffered import seekable_stream, BUFFER_SIZE
+from messytables.text import UTF8Recoder, to_unicode_or_bust
 from messytables.core import RowSet, TableSet, Cell
 from messytables.error import ReadError
 
-
-class UTF8Recoder:
-    """Iterator that reads an encoded stream and re-encodes it to UTF-8."""
-
-    # maps between chardet encoding and codecs bom keys
-    BOM_MAPPING = {
-        'utf-16le': 'BOM_UTF16_LE',
-        'utf-16be': 'BOM_UTF16_BE',
-        'utf-32le': 'BOM_UTF32_LE',
-        'utf-32be': 'BOM_UTF32_BE',
-        'utf-8': 'BOM_UTF8',
-        'utf-8-sig': 'BOM_UTF8',
-
-    }
-
-    def __init__(self, f, encoding):
-        sample = f.read(2000)
-        if not encoding:
-            results = chardet.detect(sample)
-            encoding = results['encoding']
-            if not encoding:
-                # Don't break, just try and load the data with
-                # a semi-sane encoding
-                encoding = 'utf-8'
-        f.seek(0)
-        self.reader = codecs.getreader(encoding)(f, 'ignore')
-
-        # The reader only skips a BOM if the encoding isn't explicit about its
-        # endianness (i.e. if encoding is UTF-16 a BOM is handled properly
-        # and taken out, but if encoding is UTF-16LE a BOM is ignored).
-        # However, if chardet sees a BOM it returns an encoding with the
-        # endianness explicit, which results in the codecs stream leaving the
-        # BOM in the stream. This is ridiculously dumb. For UTF-{16,32}{LE,BE}
-        # encodings, check for a BOM and remove it if it's there.
-        if encoding.lower() in self.BOM_MAPPING:
-            bom = getattr(codecs, self.BOM_MAPPING[encoding.lower()], None)
-            if bom:
-                # Try to read the BOM, which is a byte sequence, from
-                # the underlying stream. If all characters match, then
-                # go on. Otherwise when a character doesn't match, seek
-                # the stream back to the beginning and go on.
-                for c in bom:
-                    if f.read(1) != c:
-                        f.seek(0)
-                        break
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        line = self.reader.readline()
-        if not line or line == '\0':
-            raise StopIteration
-        result = line.encode("utf-8")
-        return result
-
-    next = __next__
-
-
-def to_unicode_or_bust(obj, encoding='utf-8'):
-    if isinstance(obj, binary_type):
-        obj = text_type(obj, encoding)
-    return obj
+DELIMITERS = ['\t', ',', ';', '|']
 
 
 class CSVTableSet(TableSet):
@@ -91,7 +28,7 @@ class CSVTableSet(TableSet):
         self.skipinitialspace = skipinitialspace
 
     def make_tables(self):
-        """ Return the actual CSV table. """
+        """Return the actual CSV table."""
         return [CSVRowSet(self.name, self.fileobj,
                           delimiter=self.delimiter,
                           quotechar=self.quotechar,
@@ -112,12 +49,12 @@ class CSVRowSet(RowSet):
                  encoding='utf-8', window=None, doublequote=None,
                  lineterminator=None, skipinitialspace=None):
         self.name = name
-        seekable_fileobj = seekable_stream(fileobj)
-        self.fileobj = UTF8Recoder(seekable_fileobj, encoding)
+        self.fh = seekable_stream(fileobj)
+        self.fileobj = UTF8Recoder(self.fh, encoding)
 
         def fake_ilines(fobj):
             for row in fobj:
-                    yield row.decode('utf-8')
+                yield row.decode('utf-8')
         self.lines = fake_ilines(self.fileobj)
         self._sample = []
         self.delimiter = delimiter
